+++ packages/assignment-7/src/main.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
-import {App} from './App.tsx'
+import App from './App.tsx'
 import { ChakraProvider } from '@chakra-ui/react'
 
 async function prepare() {
