+++ packages/assignment-7/src/utils/utilsFunc.ts
@@ -1,0 +1,57 @@
+export const notificationOptions = [
+    { value: 1, label: "1분 전" },
+    { value: 10, label: "10분 전" },
+    { value: 60, label: "1시간 전" },
+    { value: 120, label: "2시간 전" },
+    { value: 1440, label: "1일 전" },
+];
+
+export const getDaysInMonth = (year: number, month: number) => {
+    return new Date(year, month, 0).getDate();
+};
+
+export const getWeekDates = (date: Date) => {
+    const day = date.getDay();
+    const diff = date.getDate() - day + (day === 0 ? -6 : 1);
+    const monday = new Date(date.setDate(diff));
+    const weekDates = [];
+    for (let i = 0; i < 7; i++) {
+        const nextDate = new Date(monday);
+        nextDate.setDate(monday.getDate() + i);
+        weekDates.push(nextDate);
+    }
+    return weekDates;
+};
+
+export const formatWeek = (date: Date): string => {
+    const year = date.getFullYear();
+    const month = date.getMonth() + 1;
+    const weekNumber = Math.ceil(date.getDate() / 7);
+    return `${year}년 ${month}월 ${weekNumber}주`;
+};
+
+export const formatMonth = (date: Date): string => {
+    const year = date.getFullYear();
+    const month = date.getMonth() + 1;
+    return `${year}년 ${month}월`;
+};
+
+export const fetchHolidays = (year: number, month: number) => {
+    return {
+        "2024-01-01": "신정",
+        "2024-02-09": "설날",
+    };
+};
+
+export const parseDateTime = (date: string, time: string): Date => {
+    return new Date(`${date}T${time}`);
+};
+
+export const isOverlapping = (event1: Event, event2: Event): boolean => {
+    const start1 = parseDateTime(event1.date, event1.startTime);
+    const end1 = parseDateTime(event1.date, event1.endTime);
+    const start2 = parseDateTime(event2.date, event2.startTime);
+    const end2 = parseDateTime(event2.date, event2.endTime);
+
+    return start1 < end2 && start2 < end1;
+};
